fleshed out "actor.usr" interface, and stream behavior adjustment
git-svn-id: svn://svn.open-ils.org/ILS/trunk@497 dcc99617-32d9-48b4-a31d-7c20da2025e4

+++ Open-ILS/src/sql/Postgres/090.schema.action.sql
@@ -27,7 +27,7 @@ CREATE TABLE action.survey_answer (
 
 CREATE TABLE action.survey_response (
 	id		BIGSERIAL	PRIMARY KEY,
-	usr		BIGINT		NOT NULL, -- REFERENCES actor.usr
+	usr		INT		NOT NULL, -- REFERENCES actor.usr
 	survey		INT		NOT NULL REFERENCES action.survey,
 	question	INT		NOT NULL REFERENCES action.survey_question,
 	answer		INT		NOT NULL REFERENCES action.survey_answer,
